Fix for crash in AXIsolatedObject::relativeFrame. https://bugs.webkit.org/show_bug.cgi?id=213363
Reviewed by Chris Fleizach.

Covered by existing testss.

Between the time an isolated object dispatches the method to the main
thread and the time the lambda is executed, the isolated object is
detached and hence its object ID becomes invalid. Thus, trying to get
the associated AX object results in an assert/crash.

* accessibility/isolatedtree/AXIsolatedObject.h:


Identifier: 226451@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263573 268f45cc-cd09-0410-ab3c-d52691b4dbfc