pointerevents/ios/touch-action-none-on-iframe.html times out https://bugs.webkit.org/show_bug.cgi?id=213846 <rdar://problem/60393092>
Reviewed by Antti Koivisto.

The test pointerevents/ios/touch-action-none-on-iframe.html was mistakenly assuming that setting `touch-action: none` on
an <iframe> should prevent scrolling on the outer document with a swipe gesture over the inner document. We're renaming
this test as pointerevents/ios/touch-action-none-on-iframe-outer.html and modifying it to expect scrolling to occur, and
also make it a promise_test to avoid further timeouts due to failures in an async_test.

We also add a new test pointerevents/ios/touch-action-none-on-iframe-inner.html which sets `touch-action: none` on the
inner document which should indeed prevent scrolling of both the inner and outer documents.

* pointerevents/ios/touch-action-none-on-iframe-inner-expected.txt: Added.
* pointerevents/ios/touch-action-none-on-iframe-inner.html: Added.
* pointerevents/ios/touch-action-none-on-iframe-outer-expected.txt: Renamed from LayoutTests/pointerevents/ios/touch-action-none-on-iframe-expected.txt.
* pointerevents/ios/touch-action-none-on-iframe-outer.html: Renamed from LayoutTests/pointerevents/ios/touch-action-none-on-iframe.html.


Identifier: 226647@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263804 268f45cc-cd09-0410-ab3c-d52691b4dbfc