[iOS] Implement support for UIWKDocumentRequestSpatialAndCurrentSelection https://bugs.webkit.org/show_bug.cgi?id=213704 <rdar://problem/59738878>
Reviewed by Wenson Hsieh.

Source/WebKit:

Implement support for the new request type UIWKDocumentRequestSpatialAndCurrentSelection.
Requests of this type return the contents inside the range that covers both the specified
rect and the current selection, if there is one.

The flag UIWKDocumentRequestSpatialAndCurrentSelection was added in <rdar://problem/64867540>.

* Platform/spi/ios/UIKitSPI.h: Same thing I did in TestWebKitAPI/ios/UIKitSPI.h. See Tools ChangeLog for why.
* Shared/DocumentEditingContext.h:
* UIProcess/ios/WKContentViewInteraction.mm:
(toWebDocumentRequestOptions):
* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::requestDocumentEditingContext):

Tools:

Add some tests.

* TestWebKitAPI/Tests/WebKitCocoa/DocumentEditingContext.mm:
(TEST):
* TestWebKitAPI/ios/UIKitSPI.h: Define the enumerator using a macro just like what was done for
UIWKDocumentRequestMarkedTextRects. The reason for the macro instead of modifying the enumeration
definition in this file is to 1) avoid breaking the build for Apple engineers that haven't picked
up <rdar://problem/64867540> and to 2) avoid breaking the OpenSource build.

Identifier: 226656@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263813 268f45cc-cd09-0410-ab3c-d52691b4dbfc