[GTK] Layout test media/context-menu-actions.html is failing https://bugs.webkit.org/show_bug.cgi?id=183611
Reviewed by Xabier Rodriguez-Calvar.

Changing to fullscreen is an async operation, so we now use testExpectedEventually() to wait for
the final value.
Simply setting an event handler on video.onwebkitfullscreenchange wouldn't have been enough, since the
event is triggered before the fullscreen operation has been completed and checking for
video.webkitDisplayingFullscreen == true would still fail by then.
Using the 'webkitbeginfullscreen' event isn't possible either, because it's never triggered when
FULLSCREEN_API is enabled.

* media/context-menu-actions.html: Use asynchronous testing.
* media/video-test.js:
(testExpectedEventually): Add an optional "timeout" parameter to avoid a test timeout (all the output
would be lost) in case we wait forever for the comparison to evaluate to the expected value.
(reportExpected): Add an optional "explanation" parameter for extra reporting.
* platform/gtk-wayland/TestExpectations: Unskipped the test.
* platform/gtk/TestExpectations: Ditto.


Identifier: 226676@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263835 268f45cc-cd09-0410-ab3c-d52691b4dbfc