[GTK4] Enable threaded rendering by default https://bugs.webkit.org/show_bug.cgi?id=213883
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2020-07-03
Reviewed by Adrian Perez de Castro.

When building with GTK4 we no longer use the cairo context to render directly, so we can always enable the
threaded rendering. In WPE port threaded rendering is only enabled when env var WEBKIT_NICOSIA_PAINTING_THREADS
is present. This patch changes a bit the meaning of the env var, so that when 0 is passed the feature is
disabled too, instead of setting the number of threads to 4.

* platform/graphics/nicosia/NicosiaPaintingEngine.cpp:
(Nicosia::PaintingEngine::create):

Identifier: 226716@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263888 268f45cc-cd09-0410-ab3c-d52691b4dbfc