[GTK] WebProcess hangs when browsing GitHub https://bugs.webkit.org/show_bug.cgi?id=213970
Reviewed by Sergio Villar Senin.

Use a lower priority for LayerFlushTimer and DisplayRefreshMonitorTimer. We were using a very high priority for
drawing with the idea of keeping a good rendering performance without being affected by other timers. The
problem is that animations can be controlled by timers, so we need to ensure that MainThreadSharedTimer has
higher priority than drawing.

* wtf/glib/RunLoopSourcePriority.h: Use 110 for LayerFlushTimer and DisplayRefreshMonitorTimer.

Identifier: 226833@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264015 268f45cc-cd09-0410-ab3c-d52691b4dbfc