REGRESSION(r263836): [GStreamer] Debug ASSERT hits https://bugs.webkit.org/show_bug.cgi?id=214047
Patch by Philippe Normand <pnormand@igalia.com> on 2020-07-08
Reviewed by Xabier Rodriguez-Calvar.

Reduce the element dependency on RefPtr<MediaStreamTrackPrivate> which is not MT-safe. The
data we need from it are its ID and the contents required to build a GstTagList. So we now
create this tag-list upfront and keep track of the trackID as well.

* platform/mediastream/gstreamer/GStreamerMediaStreamSource.cpp:
(webkitMediaStreamSrcAddPad):
(ProbeData::ProbeData):
(webkitMediaStreamSrcPadProbeCb):
(webkitMediaStreamSrcSetupSrc):

Identifier: 226872@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264062 268f45cc-cd09-0410-ab3c-d52691b4dbfc