[Flatpak SDK] Pre-start sccache server https://bugs.webkit.org/show_bug.cgi?id=213973
Patch by Philippe Normand <pnormand@igalia.com> on 2020-07-08
Reviewed by Žan Doberšek.

Pre-start the sccache-dist server when the flatpak sandbox is being invoked for a build.
Usually the first sccache client process would do it but it seems to be racy when a large
number of clients is being started simultaneously.

* flatpak/flatpakutils.py:
(WebkitFlatpak.run_in_sandbox):

Identifier: 226880@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264096 268f45cc-cd09-0410-ab3c-d52691b4dbfc