REGRESSION(r217248): ASSERT_NOT_REACHED() in toWebFrameLoadType() https://bugs.webkit.org/show_bug.cgi?id=214149
Reviewed by Joseph Pecoraro.

ASSERT_NOT_REACHED() is hit when the type is FrameLoadType::ReloadExpiredOnly. This happens
when remote inspection triggers a reload via InspectorPageAgent::reload() and the
'revalidateAllResources' optional argument is false or not specified.

* WebView/WebFrame.mm:
(toWebFrameLoadType):
Remove the assert and add a note explaining why this load type that does not
naturally occur in WK1 may nonetheless be triggered while remote inspecting.


Identifier: 226964@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264195 268f45cc-cd09-0410-ab3c-d52691b4dbfc