[JSC] We should keep unaligned access feature in certain architectures in macro-assembler https://bugs.webkit.org/show_bug.cgi?id=214243
Reviewed by Darin Adler.

We introduced the assertion in r263049, but this assertion crashes in testb3 Debug build.
testb3 actually tests unaligned access feature since ARM64 and x64 allow it. And unaligned access is useful
for Yarr etc., so we want to keep unaligned access feature if architecture allows it. We should make this
assertion effective only if CPU(NEEDS_ALIGNED_ACCESS) is true.

* assembler/MacroAssembler.h:
(JSC::MacroAssembler::loadPtr):


Identifier: 227052@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264287 268f45cc-cd09-0410-ab3c-d52691b4dbfc